Make ignoring relative paths possible
